+++ internal/cli/api_test.go
@@ -17,27 +17,161 @@ func makeHTTPServer(t testing.TB, cb func(c *agent.Config)) *agent.TestAgent {
 	return agent.NewTestAgent(t, t.Name(), cb)
 }
 
+// makeHTTPServer returns a test server whose logs will be written to
+// the passed writer. If the writer is nil, the logs are written to stderr.
+func makeACLEnabledHTTPServer(t testing.TB, cb func(c *agent.Config)) *agent.TestAgent {
+	aclCB := func(c *agent.Config) {
+		cb(c)
+		ACLEnabled()(c)
+	}
+	srv := agent.NewTestAgent(t, t.Name(), aclCB)
+	testutil.WaitForLeader(t, srv.Agent.RPC)
+	c, err := NewTestClient(srv)
+	require.NoError(t, err)
+	tResp, _, err := c.ACL().Bootstrap(c.WriteOpts().Ctx())
+	require.NoError(t, err)
+
+	// Store the bootstrap ACL secret in the TestServer's client meta map
+	// so that it is easily accessible from tests.
+	srv.Config.Client.Meta = make(map[string]string, 1)
+	srv.Config.Client.Meta["token"] = tResp.GetSecretID()
+
+	return srv
+}
+
+// httpTestWithACLParallel generates a ACL enabled single node cluster and
+// automatically enables test parallelism.
+func httpTestWithACLParallel(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent)) {
+	httpTestWithACLs(t, cb, f, true)
+}
+
+// httpTestWithACL generates an ACL enabled single node cluster, without
+// automatically enabling test parallelism. This is necessary for any test that
+// uses the environment.
+func httpTestWithACL(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent)) {
+	httpTestWithACLs(t, cb, f, false)
+}
+
+// Use httpTestWithACLParallel or httpTestWithACL instead.
+func httpTestWithACLs(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent), parallel bool) {
+	if parallel {
+		t.Parallel()
+	}
+	s := makeACLEnabledHTTPServer(t, cb)
+	defer s.Shutdown()
+	// Leadership is waited for in makeACLEnabledHTTPServer
+	f(s)
+}
+
+// httpTest generates a non-ACL enabled single node cluster, without automatically
+// enabling test parallelism. This is necessary for any test that uses the
+// environment.
 func httpTest(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent)) {
+	httpTests(t, cb, f, false)
+}
+
+// httpTest generates a non-ACL enabled single node cluster and automatically
+// enables test parallelism.
+func httpTestParallel(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent)) {
 	// Since any test that uses httpTest has a distinct TestAgent, we can
 	// automatically parallelize these tests
-	t.Parallel()
+	httpTests(t, cb, f, true)
+}
+
+// Use httpTestParallel or httpTest instead.
+func httpTests(t *testing.T, cb func(c *agent.Config), f func(srv *agent.TestAgent), parallel bool) {
+	if parallel {
+		t.Parallel()
+	}
 	s := makeHTTPServer(t, cb)
 	defer s.Shutdown()
 	testutil.WaitForLeader(t, s.Agent.RPC)
 	f(s)
 }
 
-func NewTestClient(testAgent *agent.TestAgent) (*v1.Client, error) {
-	c, err := v1.NewClient(v1.WithAddress(testAgent.HTTPAddr()))
+// httpTestMultiRegionCluster generates a multi-region two node cluster and
+// automatically enables test parallelism. This will panic on test which use
+// the environment. For those, use httpTestMultiRegionCluster instead.
+func httpTestMultiRegionClusterParallel(t *testing.T, cb1, cb2 func(c *agent.Config), f func(s1 *agent.TestAgent, s2 *agent.TestAgent)) {
+	httpTestMultiRegionClusters(t, cb1, cb2, f, true)
+}
+
+// httpTestMultiRegionCluster generates a multi-region two node cluster, without
+// automatically enabling test parallelism. This is necessary for any test that
+// uses the environment.
+func httpTestMultiRegionCluster(t *testing.T, cb1, cb2 func(c *agent.Config), f func(s1 *agent.TestAgent, s2 *agent.TestAgent)) {
+	httpTestMultiRegionClusters(t, cb1, cb2, f, false)
+}
+
+// Use httpTestMultiRegionClusterParallel or httpTestMultiRegionCluster instead.
+func httpTestMultiRegionClusters(t *testing.T, cb1, cb2 func(c *agent.Config), f func(s1 *agent.TestAgent, s2 *agent.TestAgent), parallel bool) {
+	if parallel {
+		t.Parallel()
+	}
+	s1, s2 := makeMultiRegionCluster(t, cb1, cb2)
+	defer func(s1, s2 *agent.TestAgent) {
+		s1.Shutdown()
+		s2.Shutdown()
+	}(s1, s2)
+
+	testutil.WaitForLeader(t, s1.Agent.RPC)
+	testutil.WaitForLeader(t, s2.Agent.RPC)
+
+	f(s1, s2)
+}
+
+func makeMultiRegionCluster(t testing.TB, cb1, cb2 func(c *agent.Config)) (s1, s2 *agent.TestAgent) {
+	s1 = agent.NewTestAgent(t, fmt.Sprintf("%s-%s", t.Name(), "-s1"), cb1)
+	s2 = agent.NewTestAgent(t, fmt.Sprintf("%s-%s", t.Name(), "-s2"), cb2)
+
+	join(s1, s2)
+
+	return s1, s2
+}
+
+// join takes the first node and joins all of the remaining nodes provided to
+// it. When no nodes or 1 node is passed in, this function returns immediately.
+func join(nodes ...*agent.TestAgent) {
+	if len(nodes) == 0 || len(nodes) == 1 {
+		return
+	}
+	first := nodes[0]
+	var addrs = make([]string, len(nodes)-1)
+	for i, node := range nodes[1:] {
+		member := node.Agent.Server().LocalMember()
+		addrs[i] = fmt.Sprintf("%s:%d", member.Addr, member.Port)
+	}
+	count, err := first.Client().Agent().Join(addrs...)
+	require.NoError(first.T, err)
+	require.Equal(first.T, len(nodes)-1, count)
+}
+
+func NewTestClient(testAgent *agent.TestAgent, opts ...v1.ClientOption) (*v1.Client, error) {
+	maybeTokenFn := func() func(*v1.Client) { return func(*v1.Client) {} }
+
+	if token := testAgent.Config.Client.Meta["token"]; token != "" {
+		testAgent.T.Logf("building test client with token %q", token)
+		maybeTokenFn = func() func(*v1.Client) {
+			fmt.Printf("applying token %q\n", token)
+			return v1.WithToken(token)
+		}
+	}
+
+	// Push the address and possible token to the end of the options list since
+	// they are applied in order to the client config.
+	opts = append(opts, v1.WithAddress(testAgent.HTTPAddr()), maybeTokenFn())
+	c, err := v1.NewClient(opts...)
+
 	if err != nil {
 		return nil, err
 	}
 
+	testAgent.T.Log(FormatAPIClientConfig(c))
 	return c, nil
 }
 
 func Test_TestAgent_Simple(t *testing.T) {
-	httpTest(t, WithDefaultConfig(), func(s *agent.TestAgent) {
+	httpTestParallel(t, WithDefaultConfig(), func(s *agent.TestAgent) {
 		client, err := NewTestClient(s)
 		require.NoError(t, err)
 
@@ -54,7 +188,7 @@ func Test_TestAgent_Simple(t *testing.T) {
 }
 
 func Test_TestAgent_TLSEnabled(t *testing.T) {
-	httpTest(t,
+	httpTestParallel(t,
 		WithAgentConfig(
 			TLSEnabled(),
 			LogLevel(testLogLevel),
@@ -83,6 +217,36 @@ func Test_TestAgent_TLSEnabled(t *testing.T) {
 	)
 }
 
+func Test_MultiRegionCluster(t *testing.T) {
+	httpTestMultiRegionCluster(t,
+		WithAgentConfig(LogLevel(testLogLevel), Region("rA")),
+		WithAgentConfig(LogLevel(testLogLevel), Region("rB")),
+		func(s1, s2 *agent.TestAgent) {
+			c1, err := NewTestClient(s1)
+			require.NoError(t, err)
+			r, err := c1.Regions().GetRegions(c1.QueryOpts().Ctx())
+			require.NoError(t, err)
+
+			require.ElementsMatch(t, []string{"rA", "rB"}, *r)
+		},
+	)
+}
+
+func Test_MultiRegionClusterParallel(t *testing.T) {
+	httpTestMultiRegionClusterParallel(t,
+		WithAgentConfig(LogLevel(testLogLevel), Region("rA")),
+		WithAgentConfig(LogLevel(testLogLevel), Region("rB")),
+		func(s1, s2 *agent.TestAgent) {
+			c1, err := NewTestClient(s1)
+			require.NoError(t, err)
+			r, err := c1.Regions().GetRegions(c1.QueryOpts().Ctx())
+			require.NoError(t, err)
+
+			require.ElementsMatch(t, []string{"rA", "rB"}, *r)
+		},
+	)
+}
+
 // TestAgent configuration helpers
 
 // AgentOption is a functional option used as an argument to `WithAgentConfig`
@@ -101,6 +265,13 @@ func WithAgentConfig(opts ...AgentOption) func(*agent.Config) {
 		for _, opt := range opts {
 			opt(c)
 		}
+	}
+}
+
+// Region is an AgentOption used to control the log level of the TestServer
+func Region(name string) AgentOption {
+	return func(c *agent.Config) {
+		c.Region = name
 	}
 }